pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twm"
version = "0.1.0"
description = "Terminal Window Management tool for macOS Terminal.app"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "TWM Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Environment :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "PyObjC>=9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
twm = "twm.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools]
packages = ["twm"]
